+++ nix-tools.cabal
@@ -34,7 +34,6 @@ library
                      , cryptohash-sha256
                      , base16-bytestring
                      , hpack
-                     , prettyprinter
   hs-source-dirs:      lib
   default-language:    Haskell2010
 
@@ -74,9 +73,15 @@ executable hashes-to-nix
 executable plan-to-nix
   ghc-options:         -Wall
   main-is:             Main.hs
+  other-modules:       Plan2Nix
+                     , Plan2Nix.Cache
+                     , Plan2Nix.CLI
+                     , Plan2Nix.Project
+                     , Plan2Nix.Plan
   build-depends:       base >=4.11 && <4.13
                      , nix-tools
                      , hnix >= 0.5.0
+                     , Cabal >= 2.4
                      , text
                      , hpack
                      , unordered-containers
@@ -84,6 +89,13 @@ executable plan-to-nix
                      , aeson
                      , microlens
                      , microlens-aeson
+                     , optparse-applicative
+                     , prettyprinter
+                     , filepath
+                     , directory
+                     , bytestring
+                     , transformers
+                     , extra
   hs-source-dirs:      plan2nix
   default-language:    Haskell2010
